Update README.md
Expanded the database developed in a previous course to create an English query for non-technical users, to visualize the PL/SQL code used to design the database, and to show the output. Wrote 12 blocks of code ranging from 71 lines to 193 lines for each part of the project. 

Developed a DataRobot to predict where the price range would fall on clean title vehicles or insurance salvage automobiles.